Add display_battery_graph_mode

+++ firmware/atom_s3_i2c_display/lib/atom_s3_lcd/atom_s3_lcd.cpp
@@ -137,7 +137,7 @@ void AtomS3LCD::resetLcdData() {
   mode_changed = true;
 }
 
-void AtomS3LCD::setTextSize(int x) {
+void AtomS3LCD::setTextSize(float x) {
   lcd.setTextSize(x);
 }
